:bug: Fix: workflow_run pull_requests id/number types (#174)

+++ githubkit/versions/ghec_v2022_11_28/models/group_0838.py
@@ -410,8 +410,8 @@ class WebhookWorkflowRunRequestedPropWorkflowRunPropPullRequestsItems(GitHubMode
     head: WebhookWorkflowRunRequestedPropWorkflowRunPropPullRequestsItemsPropHead = (
         Field()
     )
-    id: float = Field()
-    number: float = Field()
+    id: int = Field()
+    number: int = Field()
     url: str = Field()
