+++ githubkit/versions/ghec_v2022_11_28/types/group_0836.py
@@ -368,8 +368,8 @@ class WebhookWorkflowRunCompletedPropWorkflowRunPropPullRequestsItemsType(TypedD
 
     base: WebhookWorkflowRunCompletedPropWorkflowRunPropPullRequestsItemsPropBaseType
     head: WebhookWorkflowRunCompletedPropWorkflowRunPropPullRequestsItemsPropHeadType
-    id: float
-    number: float
+    id: int
+    number: int
     url: str
