+++ githubkit/versions/ghec_v2022_11_28/types/group_0837.py
@@ -366,8 +366,8 @@ class WebhookWorkflowRunInProgressPropWorkflowRunPropPullRequestsItemsType(Typed
 
     base: WebhookWorkflowRunInProgressPropWorkflowRunPropPullRequestsItemsPropBaseType
     head: WebhookWorkflowRunInProgressPropWorkflowRunPropPullRequestsItemsPropHeadType
-    id: float
-    number: float
+    id: int
+    number: int
     url: str
